refactor(repo C13): extract rerank backpressure queue to engine/repository/rerank-queue.ts
Moves the rerank queue controller singleton, acquireRerankSlot,
getRerankQueueSnapshot, getRerankThroughputSnapshot,
runWithConcurrencyLimit, getCollectionRerankSettings, parsePositiveInt,
and all the related types into a single focused module. The Map of
controllers is module-local, preserving the singleton lifecycle.

engine/repository.ts: 4733 → 4409 lines.
Build: clean. Tests: 970/970 pass.
