+++ engine/repository/index.ts
@@ -9,3 +9,4 @@
 export * from "./paths.js";
 export * from "./llm-cache.js";
 export * from "./docid.js";
+export * from "./rerank-queue.js";
